requirements.txt
#Ensures the patch version is always updated for 3.2 but it never updates to 3.3 (can introduce breaking changes)
Django>=3.2.4,<3.3 
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16